Add official contentful content delivery API client
as this will greatly simplify accessing the API.
Consideration: The client adds quite a few dependencies and it might be
worth considering whether all it's functionality and dependencies are
truly needed or if it could be replaced with a simple implementation.
However in the interest of expediency the official client will be used
for now.

merge